+++ packages/cli/command/src/controller/controller.ts
@@ -1,14 +1,18 @@
-import { loadConfig } from "@/config/index.js";
-import { runGenerateApi } from "@mash-up-web-toolkit/generate-api";
 import {
   runAutoRouting,
   type RunAutoRoutingOptions,
-} from "@mash-up-web-toolkit/auto-routing";
-import { initConfig as initializeConfig } from "@mash-up-web-toolkit/generate-config";
-import { runGenerateSvgConfig } from "@mash-up-web-toolkit/generate-svg-config";
+} from '@mash-up-web-toolkit/auto-routing';
+import { runGenerateApi } from '@mash-up-web-toolkit/generate-api';
+import { initApiInstanceConfig } from '@mash-up-web-toolkit/generate-api-config';
+import { initConfig as initializeConfig } from '@mash-up-web-toolkit/generate-config';
+import { runGenerateSvgConfig } from '@mash-up-web-toolkit/generate-svg-config';
+
+import { loadConfig } from '@/config/index.js';
+
 export type GenApiParams = {
-  httpClientType: "fetch" | "axios";
+  httpClientType: 'fetch' | 'axios';
 };
+
 export class Controller {
   constructor() {}
 
@@ -18,7 +22,7 @@ export class Controller {
 
   async genApi(params: GenApiParams) {
     const result = await loadConfig();
-    const genApiConfig = result["gen:api"];
+    const genApiConfig = result['gen:api'];
     await runGenerateApi({
       httpClientType: params.httpClientType,
       ...genApiConfig,
@@ -31,5 +35,9 @@ export class Controller {
 
   async initConfig() {
     initializeConfig();
+  }
+
+  async initApiConfig(params: GenApiParams) {
+    initApiInstanceConfig(params.httpClientType);
   }
 }