+++ packages/cli/generate-api/src/playground/playground.ts
@@ -1,19 +1,22 @@
 import path from "node:path";
 import { generateApi } from "swagger-typescript-api";
 import { generateConfig } from "@/configs/generate-config.js";
+import { writeGeneratedApi } from "@/generator/generator.js";
 
 async function playground() {
   const result = await generateApi({
     templates: path.resolve(
       process.cwd(),
-      "./src/templates/swagger/custom-templates"
+      "./src/templates/swagger/custom-axios"
     ),
     modular: true,
-    moduleNameFirstTag: false,
+    // moduleNameFirstTag: true,
     typeSuffix: "Type",
     // output: path.resolve(process.cwd(), "./src/generated"),
     url: "https://petstore.swagger.io/v2/swagger.json",
-    httpClientType: "fetch",
+    // url: "https://api.doongdoong.org/v3/api-docs",
+    // httpClientType: "fetch",
+    httpClientType: "axios",
     hooks: {
       onPrepareConfig: (currentConfiguration) => {
         return {
@@ -26,10 +29,10 @@ async function playground() {
       },
     },
   });
-
   return result;
 }
 
 const result = await playground();
+await writeGeneratedApi(result, path.resolve(process.cwd(), "./src/generated"));
 
-console.log("#result", result);
+// console.log("#result", result);