+++ packages/cli/generate-api/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @mash-up-web-toolkit/generate-api
 
+## 0.0.15
+
+### Patch Changes
+
+- api 제너레이터 moduleNameFirstTag 속성 추가
+
+## 0.0.14
+
+### Patch Changes
+
+- Update package.json
+
+## 0.0.13
+
+### Patch Changes
+
+- Axios Template 수정
+
+## 0.0.12
+
+### Patch Changes
+
+- writeGeneratedApi 함수에 httpClientRewrite 옵션을 추가했습니다.
+
+## 0.0.11
+
+### Patch Changes
+
+- remove comment
+
+## 0.0.10
+
+### Patch Changes
+
+- test fileContent debug
+
+## 0.0.9
+
+### Patch Changes
+
+- fix axios instance
+
+## 0.0.8
+
+### Patch Changes
+
+- OAS 3.0에서 http-client 타입생성 버그 수정
+
+## 0.0.7
+
+### Patch Changes
+
+- 6bd6308: OpenAPI 3.0.1 지원을 위한 템플릿 개선
+
 ## 0.0.6
 
 ### Patch Changes